expertgpt/backend/core/requirements.txt
pymupdf==1.22.3
langchain==0.0.247
Markdown==3.4.3
openai==0.27.6
pdf2image==1.16.3
pypdf==3.9.0
StrEnum==0.4.10
supabase==1.0.3
tiktoken==0.4.0
unstructured==0.6.5
anthropic==0.2.8
fastapi==0.95.2
python-multipart==0.0.6
python-dotenv>=1.0.0
uvicorn==0.22.0
pypandoc==1.11
docx2txt==0.8
guidance==0.0.53
python-jose==3.3.0
google_cloud_aiplatform==1.25.0
transformers==4.30.1
asyncpg==0.27.0
flake8==6.0.0
flake8-black==0.3.6
sentence_transformers>=2.0.0
sentry-sdk==1.26.0
pyright==1.1.316
resend==0.5.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.18
qdrant-client==1.4.0
httpx==0.23.3
beautifulsoup4==4.12.2
passlib==1.7.4
